requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
argcomplete==1.12.3
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscrt==0.27.6
Babel==2.9.1
certifi==2025.11.12
cffi==1.14.5
chardet==4.0.0
chevron==0.13.1
click==8.1.8
clip==1.0
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
contourpy==1.3.0
cryptography==36.0.1
cycler==0.12.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
easydict==1.13
ec2-hibinit-agent==1.0.8
einops==0.8.1
eval_type_backport==0.3.0
faiss-gpu==1.7.2
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gpg==1.23.2
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.5.2
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.4.7
libcomps==0.1.20
lightning-utilities==0.15.2
lockfile==0.12.2
MarkupSafe==1.1.1
matplotlib==3.9.4
mpmath==1.3.0
netifaces==0.10.6
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
platformdirs==4.4.0
ply==3.11
prettytable==0.7.2
progressbar2==3.52.1
prompt-toolkit==3.0.24
protobuf==6.33.1
psutil==7.1.3
pycparser==2.20
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.9.0.post0
python-utils==2.4.0
pytz==2022.7.1
PyYAML==5.4.1
rank-bm25==0.2.2
regex==2025.11.3
release-notification==1.2
requests==2.25.1
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
selinux==3.4
sentencepiece==0.2.1
sentry-sdk==2.45.0
sepolicy==3.4
setools==4.4.1
six==1.15.0
sklearn==0.0
smmap==5.0.2
support-info==1.0
sympy==1.14.0
systemd-python==235
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.1
triton==3.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wandb==0.23.0
wcwidth==0.2.5
zipp==3.23.0
